type up @covector/assemble
